More removal of INT in portable code
git-svn-id: file:///local/rexx/oorexx/oorexx.svn@1181 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ sandbox/mark/os2-types/kernel/platform/windows/VariablePool.cpp
@@ -388,10 +388,10 @@ ULONG SysVariablePool(
 /******************************************************************************/
 /* RexxExecutionLineInfo													  */
 /*                                                                            */
-/* Arguments: UINT *to retrieve currently processed line                      */
-/*            char * to retrieve currently processed modul                    */
+/* Arguments: size_t *to retrieve currently processed line                    */
+/*            char   * to retrieve currently processed modul                  */
 /******************************************************************************/
-ULONG REXXENTRY RexxExecutionLineInfo(ULONG * line, char * fname, BOOL next)
+ULONG REXXENTRY RexxExecutionLineInfo(size_t * line, char * fname, BOOL next)
 {
   if (!RexxQuery())                         /* Are we up?                     */
     return 1;                               /*   No, send nastygram.          */
